Consistent use of physical units & avoiding hard-coded arguments
In flowline_model: The previous flowline_model.ipynb used hard-coded widths in "grid point" units. If the user changed the grid point spacing map_dx, the hard-coded widths did not update (so their physical-unit values changed despite the user not changing them directly). This update adds a line for the user to define widths in physical units and automatically adjusts the width in grid units if map_dx changes.

Also revised to remove double-use of variable name "widths", introducing "wider_widths" instead, so that the notebook cells can be run multiple times and produce correct results.

In mass_balance_gradients: Some initialization aspects of mass_balance_gradients were slightly different in style from flowline_model. Changed so that students more easily see the parallel in the workflow.

Also added units to MBGs and corrected a typo.

Finally, for users not familiar with OGGM function calls, the use of unlabelled arguments could be confusing. Updated so that the user adjusts a single named variable to change ELA, and the notebook then implements that variable instead of a preset value to call OGGM mass balance functions.