Rework and clarification of the notebooks. (#51)
* Rework and clarification of the notebook. Tried to change the phrasing to be more glacier centric/educational than teaching about the model. Also removed all oggm_edu functions since these will become obsolete in the future. Some of the code for plotting is repeating a few times and could be put in a function but not really necessary.

* Minor changes on the wording of what flowlines are and which one we
pick.

* Rewrite and removal of the edu package. Quite some changes. Mainly the
removal of the response_time function of the edu package. This
functionality is now moved into a loop instead. Still gives the same
results as the old function. Code should be documented so that anyone
can follow it. Other changes include wording and phrasing and details
to some explanations.

* Fix kernel name.