Bump notebooks to new version OGGM Edu (#56)
* Some spelling errors and removal of another 3d plot - replaced with the
elevation + width profiles.

* Rename and update notebook to make use of the new edu interface. Also
updates book toc.

* Update to make use of the new edu classes.

* Updated to use new OGGM-Edu functionality and removal of some code.

* Make use of the new SurgingGlacier class from oggm edu.
This removes a lot of code which was mainly for changing the model
behaviour mid-run. Since this is all handled by oggm-edu it is no longer
needed. Instead most of the notebook deals with explaining how to set up
a surging glacier and how it compares to a normal glacier.
Also includes some small text changes.

* Make use of the new oggm_edu and update of a dead link. For the ice flux
plot, we rely on attributes of the glacier object and create a custom
plot. Having a method for creating an ice flux plot is maybe too
specific.

* OGGM-Edu bump. Changes to make use of new oggm_edu functionality.
Removed one of the plots. Also some grammar/spelling fixes.

* Update ELA to ela.

* Some minor changes
- Now import plot_glacier_graphics from `oggm_edu.funcs`.
- Import run_with_constant climate from legacy.
- Change `oggm.cfg.PARAMS["store_model_geometry]` to `True`.

* Complex mass balance gradients.

Add a glacier with a complex mass balance gradient. Mostly to show how
it works.